Rename SdlConnection to session
