+++ android/sdl_android/src/androidTest/java/com/smartdevicelink/test/SdlConnection/SdlSessionTests.java
@@ -1,6 +1,6 @@
 package com.smartdevicelink.test.SdlConnection;
 
-import com.smartdevicelink.SdlConnection.SdlSession;
+import com.smartdevicelink.session.SdlSession;
 import com.smartdevicelink.protocol.enums.SessionType;
 import com.smartdevicelink.proxy.interfaces.ISdlServiceListener;
 import com.smartdevicelink.test.streaming.MockInterfaceBroker;
@@ -12,7 +12,7 @@ import static androidx.test.platform.app.InstrumentationRegistry.getInstrumentat
 
 /**
  * This is a unit test class for the SmartDeviceLink library project class : 
- * {@link com.smartdevicelink.SdlConnection.SdlSession}
+ * {@link com.smartdevicelink.session.SdlSession}
  */
 public class SdlSessionTests extends TestCase {
