+++ android/sdl_android/src/main/java/com/smartdevicelink/managers/video/VideoStreamManager.java
@@ -42,7 +42,7 @@ import android.view.MotionEvent;
 
 import androidx.annotation.RestrictTo;
 
-import com.smartdevicelink.SdlConnection.SdlSession;
+import com.smartdevicelink.session.SdlSession;
 import com.smartdevicelink.encoder.VirtualDisplayEncoder;
 import com.smartdevicelink.managers.BaseSubManager;
 import com.smartdevicelink.managers.CompletionListener;
