+++ android/sdl_android/src/main/java/com/smartdevicelink/session/SdlSession.java
@@ -30,7 +30,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.smartdevicelink.SdlConnection;
+package com.smartdevicelink.session;
 
 import android.content.Context;
