+++ android/sdl_android/src/main/java/com/smartdevicelink/streaming/StreamPacketizer.java
@@ -33,7 +33,7 @@ package com.smartdevicelink.streaming;
 
 import androidx.annotation.RestrictTo;
 
-import com.smartdevicelink.SdlConnection.SdlSession;
+import com.smartdevicelink.session.SdlSession;
 import com.smartdevicelink.managers.CompletionListener;
 import com.smartdevicelink.protocol.ProtocolMessage;
 import com.smartdevicelink.protocol.enums.SessionType;
