+++ base/src/main/java/com/smartdevicelink/session/ISdlSessionListener.java
@@ -29,7 +29,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.smartdevicelink.SdlConnection;
+package com.smartdevicelink.session;
 
 import androidx.annotation.RestrictTo;
