+++ base/src/main/java/com/smartdevicelink/streaming/AbstractPacketizer.java
@@ -31,7 +31,7 @@
  */
 package com.smartdevicelink.streaming;
 
-import com.smartdevicelink.SdlConnection.SdlSession;
+import com.smartdevicelink.session.SdlSession;
 import com.smartdevicelink.protocol.enums.SessionType;
 import com.smartdevicelink.proxy.RPCRequest;
 import com.smartdevicelink.util.Version;
